Fixed comment and moved bypass to bindtypes.
